refactor(gemini): reorganize itinerary generator files into modular structure
Move types, agent logic, utils and route handlers into dedicated subdirectories for better maintainability. Keep all functionality identical while improving code organization.

The changes include:
- Moving types.ts to itinerary-generator/types/
- Moving agent.ts to itinerary-generator/agent/
- Moving itineraryUtils.ts to itinerary-generator/utils/
- Moving route.ts to itinerary-generator/route/
- Updating all import paths accordingly
