+++ src/app/api/gemini/itinerary-generator/types/types.ts
