+++ src/app/api/gemini/itinerary-generator/utils/itineraryUtils.ts
